- added ZABBIX agent for Win64 platforms. (Eugene)

+++ src/libs/zbxcommon/comms.c
@@ -35,12 +35,12 @@ int	comms_create_request(char *host, char *key, char *data, char *lastlogsize, c
 	memset(data_b64,0,sizeof(data_b64));
 	memset(lastlogsize_b64,0,sizeof(lastlogsize_b64));
 
-	str_base64_encode(host, host_b64, strlen(host));
-	str_base64_encode(key, key_b64, strlen(key));
-	str_base64_encode(data, data_b64, strlen(data));
+	str_base64_encode(host, host_b64, (int)strlen(host));
+	str_base64_encode(key, key_b64, (int)strlen(key));
+	str_base64_encode(data, data_b64, (int)strlen(data));
 	if(lastlogsize[0]!=0)
 	{
-		str_base64_encode(lastlogsize, lastlogsize_b64, strlen(lastlogsize));
+		str_base64_encode(lastlogsize, lastlogsize_b64, (int)strlen(lastlogsize));
 	}
 
 /*	fprintf(stderr, "Data Base64 [%s]\n", data_b64);*/
